+++ CDJSVis/filtering/filterer.py
@@ -1555,7 +1555,7 @@ class Filterer(object):
         atomCluster.resize(NVisible, refcheck=False)
         
         # store cluster indexes as scalars
-        self.scalarsDict["Cluster"] = np.asarray(atomCluster, dtype=np.float64)
+#         self.scalarsDict["Cluster"] = np.asarray(atomCluster, dtype=np.float64)
         
         # build cluster lists
         self.clusterList = []